+++ novo-elements.module.ts
@@ -39,6 +39,7 @@ import { NovoPopOverModule } from './src/elements/popover/PopOver.module';
 import { NovoLabelService } from './src/services/novo-label-service';
 import { NovoDragulaService } from './src/elements/dragula/DragulaService';
 import { ComponentUtils } from './src/utils/component-utils/ComponentUtils';
+import { FormUtils } from './src/elements/form/FormUtils';
 
 @NgModule({
     imports: [
@@ -82,7 +83,8 @@ import { ComponentUtils } from './src/utils/component-utils/ComponentUtils';
     providers: [
         { provide: ComponentUtils, useClass: ComponentUtils },
         { provide: NovoLabelService, useClass: NovoLabelService },
-        { provide: NovoDragulaService, useClass: NovoDragulaService }
+        { provide: NovoDragulaService, useClass: NovoDragulaService },
+        { provide: FormUtils, useClass: FormUtils }
     ]
 })
 export class NovoElementsModule {